Convert project to <Draggable> and <DraggableCore> and ES6.
<DraggableCore> is a new component with very minimal state that handles the bulk
of what <Draggable> used to handle. It is intended for use in more advanced
applications. Events thrown by <DraggableCore> include a position with simple
deltas, and it is up to the application to translate the <DraggableCore> object
based on those deltas.

<Draggable> still maintains internal state based on position and should function
identically to previous.

Bugs: Grid snapping is broken because deltas don't accumulate.