Only apply user-select hack if drag is started
Previously the user-select hack is applied whenever mouseDown on an
element matching the selector, with the right kind of click.

But if the onStart method returns false then the drag is aborted, but
the user-select hack is still applied. Because the drag is not running
the hack is not unapplied on mouseUp. This leads to a whole series of
";user-select: none;; user-select: none;; ..." appearing on the body tag
and the inability to select text on the page until a valid drag happens.

This commit moves the hack application to after the last short circuit
to ensure dragging is actually happening.